[NOD-142] Convert Height and ChainHeight to uint64 (#269)
* [NOD-142] Updated util.FastLog2Floor to work on uint64

* [NOD-142] Convert height and chainHeight to uint64

* [NOD-142] A couple fixes in comments of TestFastLog2Floor

* [NOD-142] Make spendableOutOffset uint64 too

+++ blockdag/blocknode_test.go
@@ -26,7 +26,7 @@ func TestChainHeight(t *testing.T) {
 
 	tests := []struct {
 		node                *blockNode
-		expectedChainHeight uint32
+		expectedChainHeight uint64
 	}{
 		{
 			node:                node0,
